调整Time模块，来满足编辑器模式下暂停和逐帧执行游戏的需求

+++ CPPScripts/ParticleSystem.h
@@ -48,7 +48,7 @@ namespace ZXEngine
 		vector<Particle*> particles;
 		unsigned int textureID = 0;
 		unsigned int lastUsedIndex = 0;
-		long long lastGenTime = 0;
+		float lastGenTime = 0;
 		Vector3 lastPos;
 		// 所挂载的GO当前移动方向
 		Vector3 moveDir;
